koji: Autospec creation for version 1.15.0
Update koji to the latest version.  Per the release notes, usage of
pycurl has been dropped in favor of requests.  Per the migration notes,
there are no database schema updates.

To make koji more stable on Clear as python2 is removed from the distro,
all recursive python2 requirements are explicitly stated.  This is required
because the code base for koji is still noticeably based on python2.  As the
distro made python3 the default python interpreter, the existing patch
which forced usage of the python2 interpreter with koji needed
re-created with the update to the latest version.

To further make koji more stable on Clear, builders require a workaround that
tells NSS to allow forking.  This workaround is needed because RPM is likely not
initializing NSS data structures correctly to handle forking.  Clear packages
a modern version of NSS where this forking behavior is not tolerated by default.

To tidy the spec file for koji, binaries are installed in /usr/bin as opposed to
/usr/sbin.  To make the spec file for koji more complete, adding in
missing requirements for koji to run normally so that the koji package
is more functional on its own.

Signed-off-by: George T Kramer <george.t.kramer@intel.com>