koji: Autospec creation for version 1.17.0
Build koji for python3.  This requires modifying the shebangs in the
source files, and excluding the mergerepo script because it is python2
and depends on yum.  This script is not required for normal operation.

Reorganize the patches to be put in a more logical order.  Group the
python3 patches together.

Drop all of the legacy python2 spec file requirements that were resolved
recursively and make this package a more normal package.

This removes the last dependency on yum and yum-metadata-parser and
allows these packages to be blocked.  Mash will also be blocked along
with this change because it is a deprecated python2 utility that depends
on yum.