koji: Autospec creation for version 1.19.1
Fix up patch ordering.  Want to put patches first that are required for
distribution of koji in general.  Then want to put patches that are
required for distribution of koji in Clear.  Patches are ordered with
the hope that they are last to come in, but first to go out over time.